+++ src/components/canned-responses/helpers.ts
@@ -1,3 +1,8 @@
+import { privacyFilterCannedResponses } from "../../hooks/helpers/privacy-filter-canned-responses";
+import { searchInString } from "../../hooks/helpers/search";
+import { CannedResponse } from "../../types/canned-responses";
+import { CannedResponseFilterType } from "../../types/filter-type";
+
 export function getExtractedText(text: string, maxTextLength = 150): string {
   return text.length > maxTextLength ? `${text.substring(0, maxTextLength - 3)}...` : text;
 }
@@ -5,3 +10,16 @@ export function getExtractedText(text: string, maxTextLength = 150): string {
 export function filterAvailableTags(tags: string[], availableTagNames: Set<string>): string[] {
   return availableTagNames.size > 0 ? tags.filter((tag) => availableTagNames.has(tag)) : tags;
 }
+
+export const filteredAndSortedResponses = (responses: CannedResponse[], filter: CannedResponseFilterType): CannedResponse[] =>
+  privacyFilterCannedResponses(responses, filter).sort(
+    (first, second) => second.modificationTimestamp - first.modificationTimestamp,
+  );
+
+export const searchInItem = (item: CannedResponse, search: string): boolean => {
+  if (!search) return true;
+  const { text, tags, createdBy } = item;
+  return (
+    searchInString(text, search) || searchInString(tags.join(' '), search) || searchInString(createdBy || '', search)
+  );
+};
