+++ src/hooks/helpers/search.ts
@@ -1,0 +1,3 @@
+export const searchInString = (item: string, search: string): boolean => {
+    return item.toLowerCase().includes(search.toLowerCase());
+  };
